+++ src/memic/__init__.py
@@ -4,13 +4,17 @@ from ._version import __version__
 from .client import Memic
 from .exceptions import APIError, AuthenticationError, MemicError, NotFoundError
 from .types import (
+    ColumnInfo,
     File,
     FileStatus,
     MetadataFilters,
     PageRange,
     Project,
+    ResultsContainer,
     SearchResult,
     SearchResults,
+    SearchRouting,
+    StructuredResult,
 )
 
 __all__ = [
@@ -19,13 +23,17 @@ __all__ = [
     # Client
     "Memic",
     # Types
+    "ColumnInfo",
     "File",
     "FileStatus",
     "MetadataFilters",
     "PageRange",
     "Project",
+    "ResultsContainer",
     "SearchResult",
     "SearchResults",
+    "SearchRouting",
+    "StructuredResult",
     # Exceptions
     "MemicError",
     "AuthenticationError",